rename web pkg to api
