+++ cmd/database/wire.go
@@ -5,10 +5,10 @@ package main
 import (
 	"context"
 	"github.com/google/wire"
-	"github.com/polyse/database/internal/web"
+	"github.com/polyse/database/internal/api"
 )
 
-func initWebApp(c *config) (*web.API, func(), error) {
-	wire.Build(web.NewApp, initWebAppCfg)
+func initWebApp(c *config) (*api.API, func(), error) {
+	wire.Build(api.NewApp, initWebAppCfg)
 	return nil, nil, nil
 }
